it is wayyyy too late haven't checked if the code works but i am done for now

+++ src/ESP12E/src/preset.h
@@ -1,2 +1,2 @@
-#define PROTO_SIG 159
-#define UDP_LISTEN_PORT 39368
+#define PROTO_SIG (uint8_t)159
+#define UDP_LISTEN_PORT (uint16_t)39368
